Review edits

+++ sky/tools/create_xcframework.py
@@ -3,6 +3,10 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
 import argparse
 import errno
 import os
@@ -24,20 +28,12 @@ def main():
   output_dir = os.path.abspath(args.location)
   output_xcframework = os.path.join(output_dir, '%s.xcframework' % args.name)
 
-  try:
+  if not os.path.exists(output_dir):
     os.makedirs(output_dir)
-  except OSError as e:
-    # Ignore only "file exists" errors.
-    if e.errno != errno.EEXIST:
-      raise e
 
-  try:
+  if os.path.exists(output_xcframework):
     # Remove old xcframework.
     shutil.rmtree(output_xcframework)
-  except OSError as e:
-    # Ignore only "not found" errors.
-    if e.errno != errno.ENOENT:
-      raise e
 
   # xcrun xcodebuild -create-xcframework -framework foo/baz.framework -framework bar/baz.framework -output output/
   command = ['xcrun',
@@ -45,7 +41,9 @@ def main():
     '-quiet',
     '-create-xcframework']
 
-  [command.extend(['-framework', os.path.abspath(framework)]) for framework in args.frameworks]
+  for framework in args.frameworks:
+    command.extend(['-framework', os.path.abspath(framework)])
+
   command.extend(['-output', output_xcframework])
 
   subprocess.check_call(command, stdout=open(os.devnull, 'w'))
